Add main functionality
Uses a word list to determine what is considered a valid word. Some things could be added or removed, but it works decently well for finding words that can be made with the given letters. Looks at permutations of the letters and tries to match them to the word list. Probably not as efficient as it could be, but for the game I was solving with it, the maximum number of letters given is relatively small.